+++ src/main/java/com/insurance/policy_evaluator_dsl/infrastructure/config/PolicyProperties.java
@@ -1,5 +1,7 @@
 package com.insurance.policy_evaluator_dsl.infrastructure.config;
 
+import java.math.BigDecimal;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "policy.default")
@@ -7,7 +9,7 @@ public record PolicyProperties(
         String name,
         String description,
         String eligibilityDsl,
-        Double basePremium,
+        BigDecimal basePremium,
         String variablePremiumDsl,
         String currency
 ) {
